fix: Various migration errors
